refactor(api): Rename `IncrementController` -> `CounterController`
